+++ src/physics/mesh_filament.rs
@@ -14,7 +14,7 @@ use crate::{
 };
 use crate::{
     math::{decompose_filament, dot3},
-    physics::linear_filament::vector_potential_linear_filament_scalar,
+    physics::point_source::segment::vector_potential_point_segment_scalar,
 };
 
 /// Mutual inductance from each edge in mesh 1 to each edge in mesh 2.
@@ -89,7 +89,7 @@ where
             //    with unit current in edge 1 in order to extract A/I.
             let (midpoint2, dl2) = decompose_filament(start2, end2);
             let (ax_per_amp, ay_per_amp, az_per_amp) =
-                vector_potential_linear_filament_scalar(xyzifil1, midpoint2);
+                vector_potential_point_segment_scalar(xyzifil1, midpoint2);
 
             //    Take M = dot((A/I), dL)
             let m = dot3(ax_per_amp, ay_per_amp, az_per_amp, dl2.0, dl2.1, dl2.2);