+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 4.0.0 2025-03-03
+
+### Added
+
+* !Remove `flux_density_biot_savart` backwards-compatibility alias
+* !Upgrade `flux_density_linear_filament`, `vector_potential_linear_filament`, and `body_force_density_linear_filament` functions
+  * Now handle finite wire length and finite wire thickness analytically
+  * New `wire_radius` input; A-field blends quandratically to a nonzero value at wire center, B-field blends linearly to zero
+    * Both match ideal behavior of uniform current density cylinder
+  * Old point-source segment formulations moved to `point_source::segment` module and available as `flux_density_point_segment` and `vector_potential_point_segment` functions
+* Add `field_explorer.py` example with plots comparing linear filament and point-segment calcs
+* Improve parallelism heuristics to use half of available parallelism as heuristic for physical cores
+    * Prevents oversubscription on systems with hyperthreading
+* Use x86-64-v3 reference CPU instead of manually listing instruction sets
+
 ## 3.1.0 2025-12-19
 
 ### Added